Copy b-q-i from server, trunk-pg84 branch.
The base_quality_interrogation.py script from the server should be
statically exist in the buildbot source and be called regardless of
the server's revision. It will hopefully allow to automate the
OpenERP-Test stage and parse the openerp-server's result at the slave
side.

Taken as of commit e6618b05e5ab6b.