b-q-i: blame feature: log a summary of problem/file to blame.
This is the essence of the tests: have atomic (one log entry per problem)
logs of what is failing at our tests.
The logs are parallel to any other info, at a 'bqi.blame' logging channel,
so that the caller can easily pick them. They also are machine-formatted
for easy parsing.