all: use a bzr shared repo as a proxy
When the 'proxy_url' is passed to the BzrPoller, it will instead setup
a local copy of the branches it needs to poll. Hopefully, this will work
in a shared repo, thus increasing speed.
When the 'slave_proxy_url' is passed to bzr steps, they will fetch from
that (local mirror) instead of the MS site.