buildstep: BzrSyncUp, that will fetch from slave and push to LP
It is non-trivial, because slave must never hold SSH keys nor be able
to push upstream in any way. Instead, we program the _master_ to log
into the slave's build area, fetch the merged branches and push them
upstream.
This is done in a Bash script, so that site-specific details can be
hacked without interrupting the buildbot process.