connector: cache for get_unclaimed_buildrequests()
Since the upstream buildbot framework calls this function too often,
separately for each buildername, we need to optimize and instead call
the RPC only once for all builders. We expect to have only a small
number of pending buildrequests, so it's worth to cache these only
within an interval of 10sec.