buildstep: allow bqi to have config, fix Java dbnames
Instead of the --no-bqirc, allow the b-q-i to call the
    ~/.openerp-bqirc-bbot
configuration file. There, we can place any system-specific configs
(such as the smtp conf).

Also, instead of using the group + bzr branch name, try to allocate a
short temporary name for the test database.
The names _will_ be reused in subsequent builds, but are guarded against
simultaneous use by more than one builder.