connector, all: Major refactoring for buildbot 0.8.4, openerp_libclient
This is it: the connector is now broken into Connector-component classes,
which interface the existing openerp schema to the 0.8.4 buildbot world.

At the time of this commmit, things are apparently stable, existing changes
are being built (and toasting my machine w. bzr)

Note that OpenObjectChanges are no longer being used, in favour of
Change + properties.