soft_dev.buildseries: Allow them to explicitly select slaves
The scenario is that we have a single buildbot master with multiple slaves,
different from each other (eg. i586 vs x86_64 machines, trusted/untrusted,
ones with extra capabilities or not). Then, we need to tell the series
which buildbot-slaves to use each time.

This field is optional. If not specified, all slaves will try to claim
the build.