gitpoller: when searching changes, stick to our branch
'git log old-green..new-green' would list all commits, including those
that got merged into our branch. Say, if new-green is the commit at the
"green" branch which we are usually building, it could contain commits
of the "blue" or "yellow" branches being merged into "green".

Our builders are supposed to only build "green" heads, however (eg. if
the "green" branch is the published or the trusted one). Listing a "blue"
commit would break the builds.

An idea for future implementation would be to perform the 'git log' operation
twice, get both lists, and mark then non-"green" commits with a flag that
would let ChangeFilters sort them out. Still, we have the hazard that
non-"green" commits could be duplicates that have already been registered
with other branches/changesources. Or, we would break the data integrity
if we try to set "branch = None" or "branch = 'green'" for those commits.