soft_dev.commit: Lock the table to avoid integrity errors
(separated out from previous commit, to make it more clear)

The client side, buildbot, is has an asynchronous, multitasking, design.
So, it will send all the 'submit_change()' requests in parallel. We also
use libclient, which means the parallel requests will really arrive all
concurrently.

The interesting (for other addons, too) problem is that we do a search()
to see if commits already exist, and then either create() or write(). Then,
the highly parallel scheduler of buildbot meant that we have one commit
registering itself and it's /incomplete/ parents, while some of the parents
are being submitted (aka create()'d) at the same time. Thus, the parents
will fail due to UNIQUE(hash) constraint.

We cannot afford to fail those create()s, because they contain valuable
data (of the commits). So, the IntegrityError is bad. Automatic row-wise
locking of the table (upon INSERT) won't work either, because it wouldn't
block a SELECT from querying the table (and not finding the hash).

The solution is to LOCK the table before the SELECT, thus blocking all
parallel submit_change()s or even create()s. A pythonic lock would be
worse, because it might increase chances of deadlock (holding a pythonic
lock and waiting for the implicit SQL locks).

Pending: should the ORM provide a function for that, instead of us having
this pure SQL command?