step_iface: allow module from filename, sanitized
In parsing the log lines with regexes we have 'test_name', 'msg', 'fname'
and 'module'. Log lines will be groupped by 'module.test_name', so it is
important to set the module.
The module is supposed to be a group of filenames, a logical part of the
software package. So, the (politically) correct way is to scan the log
line for _fname_, then decode the fname to module and use that.

With this commit, we allow a fallback "module = fname" rule, with some
character sanitizing (we couldn't accept space, dots or slashes).