soft_dev.commit: improve logic of author/committer
Now, the author will be split out if an explicit 'committer' is passed
in the submit_change() request.
That is, all commits have a committer (or else they wouldn't be there),
but only some of them specify the author. On the flip side, all commits
should have an author (usually the committer), but in some of them the
committer is another user.