master_poller, git: 'rescan_commits' command to fill the incomplete holes
It will only be the openerp database that knows which commits are missing,
so it must be able to perform a reverse-request to the GitPoller, in order
to scan and register them.

The command has 2 modes, 'standalone' and not-standalone. At the second
one, it would also try to scan the history since the last known branch.