steps/openerp: fix logic of random/unique dbnames
Since the buildstep is persistent and reusable, we should take care
NOT to keep the "borrowed" dbname too long. So, only render it at the
start() phase, clear the name after it has been pushed back to the pool.

This should fix the "database is already in use" kind of build errors.