steps/openerp: move the pools to step_iface.py for persistence
All python files in steps/* are dynamically loadable. Upon master
reconfiguration, they are _reloaded_, too. This is a good thing.
However, it would reset the ports_pool, dbnames_pool, which were
module-global variables. This meant that an exception would be
thrown if a build finishes after a reconfiguration. It could also
mean that dbnames would be falsely reused.

So, move them to a module which is not reloaded.