soft_dev.repos: Allow them to fork each other, multi-repo support
Thanks to Sharoon Thomas for giving me a real-life scenario for
this missing feature. :)

We have 2 projects on GitHub. Thus, two repositories. Then, we
need to mirror them to bzr, using another repo. The two git repos,
however, contain the same commits, the same trees. So, the system
must somehow know that they both belong to the same side of the
mirroring process.

In short, we used to support:
	git<->bzr
now also:
	(git+git + ...) <-> bzr