connector: API for notifications to buildbot owner
When buildbot has some trouble, it should alert its administrator
about that.
There is some cases, like the repo pollers, that won't appear
in the web status page, but the admin has to have some explicit
notification. Define this API and he will get a 'res.request'
alert in the OpenERP GUI.