soft_dev.mirrors: Lock the mirroring builds
Since mirroring is done in master side, we must make sure it that
we don't run 2 builds at the same time (multiple slaves could fool
the master and claim builds).
This had happened because of huge merges, that kept each mirroring
process running long enough for a second build to kick in. It is
dangerous, because the intermediate files would be overwritten.