lp_poller: avoid expensive search() of each branch
Every time a bzr branch was matched, a separate search_read()
would be issued to the server to locate if it existed there.
Both with fallback_search on and off, this is an expensive
operation, which led to 100% cpu on the OpenERP-F3 server.

Now, group all these calls per repository id and let the
server respond with more results.